final/requirements.txt
flask
gunicorn==20.0.4; python_version > '3.0'
gunicorn==19.10.0; python_version < '3.0'
google-cloud-vision
google-cloud-storage
google-cloud-datastore
google-cloud-translate
google-api-core
google-api-python-client
google-cloud
